float step (fixes issue #1)

+++ widgets/Spinbox.h
@@ -21,7 +21,7 @@ protected:
     double maximum;
     std::string prefix;
     std::string suffix;
-    int step;
+    double step;
     bool float_;
 
 public:
